refactor(runtime): D4 — dedup resource naming (P5)
Moves dind_certs_volume, dind_container_name, role_network_name from
crates/jackin-runtime/src/runtime/naming.rs (the old home) into
crates/jackin-runtime/src/instance/naming.rs (the new single home).
The pattern matches every other naming string in the workspace
(container_name_with_id, class_family_matches, image_name,
image_name_for_branch, role_base_image_name, short_git_sha, etc.):
one canonical owner, callers reach it through a single import.

Touches (7 files modified, 0 created/deleted):
  - crates/jackin-runtime/src/instance/naming.rs: adds the three
    pub(crate) functions (verbatim from runtime/naming.rs); the
    existing dind_certs_volume_derives_from_container_name test moved
    to instance/naming/tests.rs in the same change
  - crates/jackin-runtime/src/instance/manifest.rs: repoints the three
    call sites from crate::runtime::naming to crate::instance::naming
  - crates/jackin-runtime/src/runtime/cleanup.rs: splits the
     import block to keep the four LABEL_*_*
    constants from runtime::naming and pull the two call-site helpers
    from crate::instance::naming
  - crates/jackin-runtime/src/runtime/launch.rs: same split (LABEL_*
    constants stay in runtime::naming, dind_certs_volume moves to
    crate::instance::naming); this is the only net-line change in
    the slice (+1)
  - crates/jackin-runtime/src/runtime/launch/launch_dind.rs: repoints
    two super::super::naming calls to crate::instance::naming
  - crates/jackin-runtime/src/runtime/launch/launch_pipeline.rs:
    same two-line repoint
  - crates/jackin-runtime/src/runtime/naming.rs: removes the three
    functions; the file is now 17 lines shorter and holds only the
    image-name helpers + the LABEL_IMAGE_* constants
  - crates/jackin-runtime/src/runtime/naming/tests.rs: removes the
    now-relocated dind_certs_volume test (the same test now lives in
    instance/naming/tests.rs)
  - file-size-budget.toml: ratchet entry for launch.rs bumped 2833
    → 2834 (the new
    import line replaces a single-line use of the now-removed
     entry — net +1; future
    cleanup can drop it by re-joining the import block)

Decisions:
  - instance/naming.rs is the canonical home because every other
    container-name-derived string in the workspace (container_name_with_id,
    class_family_matches, image_name, image_name_for_branch,
    role_base_image_name, short_git_sha) already lives there. Moving
    these three keeps the layer rule: 'one concept, one home'.
  - The three functions are pub(crate) at the new home so callers in
    runtime/cleanup, runtime/launch, runtime/launch/launch_dind, and
    runtime/launch/launch_pipeline continue to compile unchanged. Each
    call site that used them from runtime::naming is repointed to
    instance::naming in this commit.

Verified:
  - cargo fmt --check clean
  - cargo clippy --workspace --all-targets --all-features --locked -- -D warnings clean
  - cargo nextest run -p jackin-runtime: 869/869 pass
  - cargo xtask lint: file-size OK at launch.rs 2834 (ratchet updated);
    test-layout OK; arch gate at 3 forbidden edges (unchanged)
  - The moved test (dind_certs_volume_derives_from_container_name) is
    under instance::naming::tests and asserts the same expected
    strings as before

Refs: plans/codebase-health-playbook.md#d4--dedup-resource-naming-p5
Refs: docs/content/docs/roadmap/codebase-health-enforcement.mdx#workstream-w1
