+++ crates/jackin-runtime/src/instance/naming.rs
@@ -141,5 +141,19 @@ fn random_instance_id() -> String {
     id
 }
 
+/// Docker volume name for the TLS client certificates shared between the
+/// `DinD` sidecar (writer) and the role container (reader).
+pub(crate) fn dind_certs_volume(container_name: &str) -> String {
+    format!("{container_name}-dind-certs")
+}
+
+pub(crate) fn dind_container_name(container_name: &str) -> String {
+    format!("{container_name}-dind")
+}
+
+pub(crate) fn role_network_name(container_name: &str) -> String {
+    format!("{container_name}-net")
+}
+
 #[cfg(test)]
 mod tests;
