+++ crates/jackin-runtime/src/runtime/launch.rs
@@ -67,7 +67,8 @@ use std::path::{Path, PathBuf};
 use super::attach::{ContainerState, reconnect_or_create_session_with_focus};
 use super::discovery::list_running_agent_names;
 use super::identity::GitIdentity;
-use super::naming::{LABEL_KEEP_AWAKE, LABEL_KIND_ROLE, LABEL_MANAGED, dind_certs_volume};
+use super::naming::{LABEL_KEEP_AWAKE, LABEL_KIND_ROLE, LABEL_MANAGED};
+use crate::instance::naming::dind_certs_volume;
 use super::universe::ExitClaim;
 use jackin_docker::docker_client::DockerApi;
