+++ crates/jackin-runtime/src/runtime/naming.rs
@@ -191,19 +191,5 @@ pub(super) fn role_base_image_name(
     tag_with_sha(repo, role_git_sha)
 }
 
-/// Docker volume name for the TLS client certificates shared between the
-/// `DinD` sidecar (writer) and the role container (reader).
-pub(crate) fn dind_certs_volume(container_name: &str) -> String {
-    format!("{container_name}-dind-certs")
-}
-
-pub(crate) fn dind_container_name(container_name: &str) -> String {
-    format!("{container_name}-dind")
-}
-
-pub(crate) fn role_network_name(container_name: &str) -> String {
-    format!("{container_name}-net")
-}
-
 #[cfg(test)]
 mod tests;
